Created Progress Bar for skills with text overlay Implement progress bar for skills with current XP and required XP Add LinearProgressIndicator with text for skill levels

+++ lib/models/skills.dart
@@ -2,8 +2,8 @@ class Skill {
   final int xp;
   final int level;
   final int maxLevel;
-  final int xpCurrent;
-  final int xpForNext;
+  final double xpCurrent;
+  final double xpForNext;
   final double progress;
   final int levelCap;
   final int uncappedLevel;
@@ -30,8 +30,8 @@ class Skill {
       xp: (json['xp'] ?? 0.0).toInt(), // Convert xp to integer
       level: json['level'] ?? 0,
       maxLevel: json['maxLevel'] ?? 0,
-      xpCurrent: (json['xpCurrent'] ?? 0.0).toInt(), // Convert xpCurrent to integer
-      xpForNext: (json['xpForNext'] ?? 0.0).toInt(), // Convert xpForNext to integer
+      xpCurrent: (json['xpCurrent'] ?? 0.0), 
+      xpForNext: (json['xpForNext'] ?? 0.0), 
       progress: json['progress'] ?? 0.0,
       levelCap: json['levelCap'] ?? 0,
       uncappedLevel: json['uncappedLevel'] ?? 0,
